requirements.txt
requests
pandas
matplotlib
schedule
python-dotenv
